Add a high level restate protocol parsing.
This commit adds a RestateDuplexStream which is a NodeJS stream
transformer. You can think about it as a Netty handler that can
transform low level ByteBuf streams to high level object streams.
This transformer parses/encodes Protobuf messages prefixied with a
Header as defined in restate's runtime header.rs and the protocol
buffers definitions.