Correctly handle replays while awaiting promises
On return, we await all promises that we haven't seen an incomplete replay for - ie, a replay where there is not a completion, like for example because we had an invocation we didn't await. We previously added the check re having already seen an incomplete replay because of deadlock in cases where we invoke ourself at the same key, but don't await it. However, we currently have a race condition where the hasBeenChecked flag can be set after we are already waiting on the Promise, and so we don't know not to wait on it.

The solution is to short circuit the Promise.all for pending promises that get the 'checked' marker added as the replay comes in. Hopefully this can all be simplified in rewrite, but for the time being this works.