Remove generic parameters from RestateContextImpl.
The generic parameters were not used by the context, but only there to be used on the reference to the StateMachine.
However, for the purpose of the Context's use of the StateMachine, they are irrelevant, so we drop them.

Note: The cleanest solution would be factoring out the relevant parts of the StateMachine (accepting user code messages) into a separate non-generic interface.
That would also remove the slightly-not-so-nice use of 'any' for the generic parameters of the reference to the state machine inside the context.