Don't report wrong API use errors through side channel.
This follows the paradigm to minimize error reporting through side channels (directly from code to state machine)
and use whereever possible the canonical way (throwing and bubbling up through user code) to reduce the surprises
when users catch exceptions.

This now allows users to violate the API, catch that error, and continue, which is ok, because it does not leave
anything in an inconsistent state.