Minor logging changes
A few small logging changes to align better with how the server interacts with the SDK in some cases. The most notable is that when we discover services we don't always "register" them (in dry-run mode) so that SDK should just log that it's announcing those services instead. Additionally. Using the word "registering" in logs might confuse users as it's not registering those services to restate server but rather just binding those services on the open port.