Catch send/sendDelayed failed promises (#273)
Send/send delayed are `void` methods on our public api surface,
but the reality is that these operations produce a promise, and this
promise can actually fail. For example due to a non determinism errors.
This commit makes sure to propagate errors happening in these promises
back to the state machine.
See the e2e test called `test_non-determinism` in the e2e repo.