Add a .catch() for the dangling promises
In Node, a rejected promises without a .catch() causes
the process to exit. We have few places in our API that create
promises internally but do not expose them back to the caller.
For example: ctx.set(), ctx.reject/resolve awakable, etc'.
These promises might get rejected, due to for example non determinism
errors.
This commit adds a .catch() call, that forwards the failure back to the
FSM.