[api] Add shared handlers
This commit adds shared handlers for virtual objects
With this commit and runtime >= 0.9, it is possible
to define handlers as shared.
A shared handler allows concurrent invocations for a given key,
but it does not allow any state modification (only retrieval)
This is useful to retrieve state keys for locked objects.