[workflow] Adapt the workflow client to the new ingress structure
Previously the workflow client wrapped the requests with a 'request'
object, and unwrapped the responses with a 'response' object.
This is not needed anymore as the ingress passes the payloads as-is.
Note that, the next step would be to use the restate `client` to wrap
the proper interaction with the workflow service.