[workflow] Various adjustments to the workflow overly impl (#346)
* [workflow] Use state service name suffix for the aux state object

* [workflow] Adapt the workflow client to the new ingress structure

Previously the workflow client wrapped the requests with a 'request'
object, and unwrapped the responses with a 'response' object.
This is not needed anymore as the ingress passes the payloads as-is.
Note that, the next step would be to use the restate `client` to wrap
the proper interaction with the workflow service.

* [workflow] Rename start -> submit

Workflow service wrapper uses the name 'submit',
while the WorkflowAPI used for internal communication uses the name
'start'. This commit aligns the names.